requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers>=4.0.0
scikit-learn>=1.2.0
sentence-transformers>=2.2.2
Pillow==10.4.0
faiss-cpu>=1.7.0
openai>=1.0.0
torch>=2.0.0
transformers>=4.30.0
python-multipart>=0.0.6
opencv-python==4.8.1.78
numpy==1.26.4
face-recognition==1.3.0
dlib==20.0.0
channels
reportlab